requirements.txt
# Core dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Common utility packages that might be useful for this project
requests>=2.31.0
pydantic>=2.0.0